\name{CorMut-package}
\alias{CorMut-package}
\alias{CorMut}
\docType{package}
\title{
Detect the correlated mutations based on selection pressure
}
\description{
CorMut provides functions for computing kaks for individual site or specific amino and detecting correlated mutation among them. Three methods are provided for detecting correlated mutation,including conditional selection pressure, mutual information and Jaccard index. the computation consists of two steps: First,the positive selection sites are detected; Second,the mutation correlation are computed among the positive selection sites. Note that the first step is optional. Meanwhile, CorMut facilitates the comparison of the correlated mutations between two conditions by the means of correlated mutation network.
}
\details{
\tabular{ll}{
Package: \tab CorMut\cr
Type: \tab Package\cr
Version: \tab 1.21.1\cr
Date: \tab 2018-02-21\cr
License: \tab GPL-2\cr
}
}
\author{
Zhenpeng Li
Maintainer: Zhenpeng Li <zpli21@gmail.com>
}
\references{
1. Chen, L., Perlina, A. & Lee, C. J. Positive selection detection in 40,000 human immunodeficiency virus (HIV) type 1 sequences automatically identifies drug resistance and positive fitness mutations in HIV protease and reverse transcriptase. Journal of virology 78,3722-3732 (2004).

2. Cover, T. M., Thomas, J. A. & Wiley, J. Elements of information theory. 6, (Wiley Online Library: 1991).
}
